feat: implement simulate_provide_liquidity for latest astroport code
