+++ Cargo.toml
@@ -15,10 +15,12 @@ osmosis = []
 osmosis-std = { git = "https://github.com/osmosis-labs/osmosis-rust" }
 cosmwasm-schema = "1.1.0"
 cosmwasm-std = {version = "1.1.0", features = ["stargate"]}
+cosmwasm-storage = "1.1.0"
 cw-asset = "2.3.0"
 cw-storage-plus = "0.16"
 cw-utils = "0.16"
 cw20-base = "0.16"
+itertools = "0.10.5"
 num-bigint = "0.4"
 num-rational = "0.4"
 num-traits = "0.2.15"
@@ -38,5 +40,5 @@ stake-cw20-external-rewards = { git = "https://github.com/DA0-DA0/dao-contracts"
 cw-utils-0_13_4 = { package = "cw-utils", version = "0.13.4" }
 
 # Astroport
-astroport-core = { package = "astroport", rev = "126d43216111df786472fe2a845c1e2fadfe4a36", git = "https://github.com/astroport-fi/astroport-core.git" }
 delegate = "0.8"
+astroport-core = { package = "astroport", git = "https://github.com/astroport-fi/astroport-core.git" }
