+++ src/implementations/astroport/astroport.rs
@@ -202,491 +202,3 @@ impl Pool for AstroportPool {
     }
 }
 
-#[cw_serde]
-pub struct AstroportStableSwapPool {
-    contract_addr: String,
-    lp_token_addr: String,
-    generator_addr: String,
-}
-
-impl AstroportStableSwapPool {
-    fn query_lp_token_supply(&self, querier: &QuerierWrapper) -> StdResult<Uint128> {
-        query_supply(querier, &self.lp_token_addr)
-    }
-
-    fn swap_native_msg(
-        &self,
-        offer_asset: &Asset,
-        ask_asset_info: &AssetInfo,
-        belief_price: Decimal,
-        recipient: Addr,
-    ) -> Result<Response, CwDexError> {
-        let asset = cw_asset_to_astro_asset(offer_asset)?;
-        let msg = PairExecMsg::Swap {
-            offer_asset: asset,
-            ask_asset_info: Some(cw_asset_info_to_astro_asset_info(ask_asset_info)?),
-            belief_price: Some(belief_price),
-            max_spread: Some(Decimal::zero()),
-            to: Some(recipient.into_string()),
-        };
-        let swap = CosmosMsg::Wasm(WasmMsg::Execute {
-            contract_addr: self.contract_addr.to_owned(),
-            msg: to_binary(&msg)?,
-            funds: vec![offer_asset.try_into()?],
-        });
-        Ok(Response::new().add_message(swap))
-    }
-
-    fn swap_cw20_msg(
-        &self,
-        offer_asset: &Asset,
-        ask_asset_info: &AssetInfo,
-        belief_price: Decimal,
-        recipient: Addr,
-    ) -> Result<Response, CwDexError> {
-        if let AssetInfoBase::Cw20(token_addr) = offer_asset.to_owned().info {
-            let swap = CosmosMsg::Wasm(WasmMsg::Execute {
-                contract_addr: token_addr.to_string(),
-                msg: to_binary(&Cw20ExecuteMsg::Send {
-                    contract: self.contract_addr.to_string(),
-                    amount: Uint128::zero(), // Should this be `offer_asset.amount`?
-                    msg: to_binary(&Cw20HookMsg::Swap {
-                        ask_asset_info: Some(cw_asset_info_to_astro_asset_info(ask_asset_info)?),
-                        belief_price: Some(belief_price),
-                        max_spread: Some(Decimal::zero()),
-                        to: Some(recipient.into_string()),
-                    })?,
-                })?,
-                funds: vec![],
-            });
-            Ok(Response::new().add_message(swap))
-        } else {
-            Err(CwDexError::InvalidInAsset {
-                a: offer_asset.to_owned(),
-            })
-        }
-    }
-
-    fn get_pool_liquidity_impl(&self, querier: &QuerierWrapper) -> StdResult<PoolResponse> {
-        let query_msg = QueryMsg::Pool {};
-        let wasm_query = WasmQuery::Smart {
-            contract_addr: self.contract_addr.to_string(),
-            msg: to_binary(&query_msg)?,
-        };
-        let query_request = QueryRequest::Wasm(wasm_query);
-        querier.query::<PoolResponse>(&query_request)
-    }
-}
-
-impl Pool for AstroportStableSwapPool {
-    fn provide_liquidity(
-        &self,
-        _deps: Deps,
-        assets: AssetList,
-        recipient: Addr,
-        slippage_tolerance: Option<Decimal>,
-    ) -> Result<Response, CwDexError> {
-        let astro_assets: Vec<AstroAsset> = AstroAssetList::try_from(assets)?.into();
-
-        let msg = PairExecMsg::ProvideLiquidity {
-            assets: astro_assets.clone(),
-            slippage_tolerance,
-            auto_stake: Some(false), // Should this be true?
-            receiver: Some(recipient.to_string()),
-        };
-        let provide_liquidity = CosmosMsg::Wasm(WasmMsg::Execute {
-            contract_addr: self.contract_addr.to_string(),
-            msg: to_binary(&msg)?,
-            funds: vec![],
-        });
-
-        let event = Event::new("apollo/cw-dex/provide_liquidity")
-            .add_attribute("pair_addr", &self.contract_addr)
-            .add_attribute("assets", format!("{:?}", astro_assets))
-            .add_attribute("recipient", recipient.to_string());
-
-        Ok(Response::new().add_message(provide_liquidity).add_event(event))
-    }
-
-    fn withdraw_liquidity(
-        &self,
-        _deps: Deps,
-        asset: Asset,
-        recipient: Addr,
-    ) -> Result<Response, CwDexError> {
-        if let AssetInfoBase::Cw20(token_addr) = &asset.info {
-            let withdraw_liquidity = CosmosMsg::Wasm(WasmMsg::Execute {
-                contract_addr: token_addr.to_string(),
-                msg: to_binary(&Cw20ExecuteMsg::Send {
-                    contract: self.contract_addr.to_string(),
-                    amount: asset.amount,
-                    msg: to_binary(&Cw20HookMsg::WithdrawLiquidity {
-                        assets: vec![],
-                    })?,
-                })?,
-                funds: vec![],
-            });
-
-            let event = Event::new("apollo/cw-dex/withdraw_liquidity")
-                .add_attribute("pair_addr", &self.contract_addr)
-                .add_attribute("asset", format!("{:?}", asset))
-                .add_attribute("token_amount", asset.amount)
-                .add_attribute("recipient", recipient.to_string());
-
-            Ok(Response::new().add_message(withdraw_liquidity).add_event(event))
-        } else {
-            Err(CwDexError::InvalidInAsset {
-                a: asset,
-            })
-        }
-    }
-
-    fn swap(
-        &self,
-        _deps: Deps,
-        offer_asset: Asset,
-        ask_asset_info: AssetInfo,
-        minimum_out_amount: Uint128,
-        recipient: Addr,
-    ) -> Result<Response, CwDexError> {
-        // Setting belief price to the minimium acceptable return and max spread to zero simplifies things
-        // Astroport will make the best possible swap that returns at least minimum_out_amount
-        let belief_price = Decimal::from_ratio(minimum_out_amount, 1u128);
-        let response = match offer_asset.info {
-            AssetInfo::Native(_) => {
-                self.swap_native_msg(&offer_asset, &ask_asset_info, belief_price, recipient)
-            }
-            AssetInfo::Cw20(_) => {
-                self.swap_cw20_msg(&offer_asset, &ask_asset_info, belief_price, recipient)
-            }
-            _ => Err(CwDexError::InvalidInAsset {
-                a: offer_asset.clone(),
-            }),
-        }?;
-        let event = Event::new("apollo/cw-dex/swap")
-            .add_attribute("pair_addr", &self.contract_addr)
-            .add_attribute("ask_asset", format!("{:?}", ask_asset_info))
-            .add_attribute("offer_asset", format!("{:?}", offer_asset.info))
-            .add_attribute("minimum_out_amount", minimum_out_amount);
-        Ok(response.add_event(event))
-    }
-
-    fn get_pool_liquidity(&self, deps: Deps) -> Result<AssetList, CwDexError> {
-        let resp = self.get_pool_liquidity_impl(&deps.querier)?;
-        Ok(AssetList::from(AstroAssetList(resp.assets)))
-    }
-
-    fn simulate_provide_liquidity(
-        &self,
-        deps: Deps,
-        env: Env,
-        assets: AssetList,
-    ) -> Result<Asset, CwDexError> {
-        let astro_assets: AstroAssetList = assets.to_owned().try_into()?;
-        let mut pools = vec![];
-        for asset in astro_assets.0.clone() {
-            pools.push(AstroStableSwapAsset {
-                info: asset.info.clone(),
-                amount: asset.info.query_pool(&deps.querier, self.contract_addr.to_string())?,
-                precision: query_asset_precision(
-                    &deps.querier,
-                    &Addr::unchecked(self.contract_addr.to_string()),
-                    asset.info,
-                )?,
-            })
-        }
-        let config =
-            query_pair_config(&deps.querier, &Addr::unchecked(self.contract_addr.to_string()))?;
-
-        if assets.len() > config.pair_info.asset_infos.len() {
-            return Err(CwDexError::Std(StdError::generic_err("Invalid number of assets. The Astroport supports at least 2 and at most 5 assets within a stable pool")));
-        }
-
-        let pools: HashMap<_, _> = config
-            .pair_info
-            .query_pools(&deps.querier, &env.contract.address)?
-            .into_iter()
-            .map(|pool| (pool.info, pool.amount))
-            .collect();
-
-        let mut non_zero_flag = false;
-
-        let mut assets_collection = astro_assets
-            .clone()
-            .0
-            .into_iter()
-            .map(|asset| {
-                // Check that at least one asset is non-zero
-                if !asset.amount.is_zero() {
-                    non_zero_flag = true;
-                }
-
-                // Get appropriate pool
-                let pool = pools.get(&asset.info).copied().ok_or_else(|| {
-                    StdError::generic_err(format!(
-                        "The asset {:?} does not belong to the pair",
-                        asset
-                    ))
-                })?;
-
-                Ok((asset, pool))
-            })
-            .collect::<Result<Vec<_>, CwDexError>>()?;
-
-        // If some assets are omitted then add them explicitly with 0 deposit
-        pools.iter().for_each(|(pool_info, pool_amount)| {
-            if !astro_assets.0.iter().any(|asset| asset.info.eq(pool_info)) {
-                assets_collection.push((
-                    AstroAsset {
-                        amount: Uint128::zero(),
-                        info: pool_info.clone(),
-                    },
-                    *pool_amount,
-                ));
-            }
-        });
-
-        if !non_zero_flag {
-            return Err(CwDexError::Std(StdError::generic_err("Event of zero transfer")));
-        }
-
-        let assets_collection = assets_collection
-            .iter()
-            .cloned()
-            .map(|(asset, pool)| {
-                let coin_precision = query_asset_precision(
-                    &deps.querier,
-                    &Addr::unchecked(self.contract_addr.to_string()),
-                    asset.to_owned().info,
-                )?;
-                Ok((
-                    asset.to_decimal_asset(coin_precision)?,
-                    Decimal256::with_precision(pool, coin_precision)?,
-                ))
-            })
-            .collect::<StdResult<Vec<(DecimalAsset, Decimal256)>>>()?;
-
-        let amp = compute_current_amp(deps, &env, config.to_owned())?;
-
-        // Initial invariant (D)
-        let old_balances = assets_collection.iter().map(|(_, pool)| *pool).collect_vec();
-        let init_d = compute_d(amp, &old_balances, config.greatest_precision)?;
-
-        // Invariant (D) after deposit added
-        let mut new_balances = assets_collection
-            .iter()
-            .map(|(deposit, pool)| Ok(pool + deposit.amount))
-            .collect::<StdResult<Vec<_>>>()?;
-
-        let deposit_d = compute_d(amp, &new_balances, config.greatest_precision)?;
-
-        let pair_info: PairInfo =
-            deps.querier.query_wasm_smart(self.contract_addr.to_string(), &QueryMsg::Pair {})?;
-
-        let n_coins = pair_info.asset_infos.len() as u8;
-
-        let total_share = query_supply(&deps.querier, &config.pair_info.liquidity_token)?;
-        let share = if total_share.is_zero() {
-            let share = deposit_d
-                .to_uint128_with_precision(config.greatest_precision)?
-                .checked_sub(MINIMUM_LIQUIDITY_AMOUNT)
-                .map_err(|_| {
-                    CwDexError::Std(StdError::generic_err(format!(
-                        "Initial liquidity must be more than {}",
-                        MINIMUM_LIQUIDITY_AMOUNT
-                    )))
-                })?;
-
-            // share cannot become zero after minimum liquidity subtraction
-            if share.is_zero() {
-                return Err(CwDexError::Std(StdError::generic_err(format!(
-                    "Initial liquidity must be more than {}",
-                    MINIMUM_LIQUIDITY_AMOUNT
-                ))));
-            }
-
-            share
-        } else {
-            // Get fee info from the factory
-            let fee_info = query_fee_info(
-                &deps.querier,
-                &config.factory_addr,
-                config.pair_info.pair_type.clone(),
-            )?;
-
-            // total_fee_rate * N_COINS / (4 * (N_COINS - 1))
-            let fee = fee_info
-                .total_fee_rate
-                .checked_mul(Decimal::from_ratio(n_coins, 4 * (n_coins - 1)))?;
-
-            let fee = Decimal256::new(fee.atomics().into());
-
-            for i in 0..n_coins as usize {
-                let ideal_balance = deposit_d.checked_multiply_ratio(old_balances[i], init_d)?;
-                let difference = if ideal_balance > new_balances[i] {
-                    ideal_balance - new_balances[i]
-                } else {
-                    new_balances[i] - ideal_balance
-                };
-                // Fee will be charged only during imbalanced provide i.e. if invariant D was changed
-                new_balances[i] -= fee.checked_mul(difference)?;
-            }
-
-            let after_fee_d = compute_d(amp, &new_balances, config.greatest_precision)?;
-
-            let share = Decimal256::with_precision(total_share, config.greatest_precision)?
-                .checked_multiply_ratio(after_fee_d.saturating_sub(init_d), init_d)?
-                .to_uint128_with_precision(config.greatest_precision)?;
-
-            if share.is_zero() {
-                return Err(CwDexError::Std(StdError::generic_err(
-                    "Insufficient amount of liquidity",
-                )));
-            }
-
-            share
-        };
-
-        let lp_token = Asset {
-            info: AssetInfoBase::Cw20(Addr::unchecked(self.lp_token_addr.to_string())),
-            amount: share,
-        };
-        Ok(lp_token)
-    }
-
-    fn simulate_withdraw_liquidity(
-        &self,
-        deps: Deps,
-        env: Env,
-        asset: Asset,
-        withdraw_assets: AssetList,
-    ) -> Result<AssetList, CwDexError> {
-        let config =
-            query_pair_config(&deps.querier, &Addr::unchecked(self.contract_addr.to_string()))?;
-        let withdraw_assets: AstroAssetList = withdraw_assets.to_owned().try_into()?;
-
-        let _burn_amount;
-        let refund_assets;
-        let amount = &asset.amount;
-        let pools = config.pair_info.query_pools(&deps.querier, self.contract_addr.to_string())?;
-        let total_share = query_supply(&deps.querier, &config.pair_info.liquidity_token)?;
-
-        if withdraw_assets.0.is_empty() {
-            _burn_amount = asset.amount;
-            refund_assets = get_share_in_assets(&pools, *amount, total_share);
-        } else {
-            // Imbalanced withdraw
-            _burn_amount =
-                imbalanced_withdraw(deps, &env, &config, *amount, &withdraw_assets.0.to_vec())?;
-
-            refund_assets = withdraw_assets.into();
-        }
-
-        Ok(refund_assets
-            .iter()
-            .map(|a| Asset {
-                info: astro_asset_info_to_cw_asset_info(&a.info),
-                amount: a.amount,
-            })
-            .collect::<Vec<Asset>>()
-            .into())
-    }
-
-    fn simulate_swap(
-        &self,
-        deps: Deps,
-        offer_asset: Asset,
-        ask_asset_info: AssetInfo,
-        _minimum_out_amount: Uint128, // Astroport doesn't allow setting max spread or belief price on simulated swaps
-        _sender: Option<String>,      // N/A for Astroport
-    ) -> StdResult<Uint128> {
-        let query_msg = QueryMsg::Simulation {
-            offer_asset: cw_asset_to_astro_asset(&offer_asset)?,
-            ask_asset_info: Some(cw_asset_info_to_astro_asset_info(&ask_asset_info)?),
-        };
-        let wasm_query = WasmQuery::Smart {
-            contract_addr: self.contract_addr.to_string(),
-            msg: to_binary(&query_msg)?,
-        };
-        let query_request = QueryRequest::Wasm(wasm_query);
-        let resp = deps.querier.query::<SimulationResponse>(&query_request)?;
-        Ok(resp.return_amount)
-    }
-}
-
-impl Staking for AstroportStableSwapPool {
-    fn stake(&self, _deps: Deps, asset: Asset, recipient: Addr) -> Result<Response, CwDexError> {
-        let stake_msg = CosmosMsg::Wasm(
-            (WasmMsg::Execute {
-                contract_addr: self.lp_token_addr.to_string(),
-                msg: to_binary(&Cw20ExecuteMsg::Send {
-                    contract: self.generator_addr.to_string(),
-                    amount: asset.amount,
-                    msg: to_binary(&GeneratorCw20HookMsg::Deposit {})?,
-                })?,
-                funds: vec![],
-            }),
-        );
-
-        let event = Event::new("apollo/cw-dex/stake")
-            .add_attribute("type", "astroport_staking")
-            .add_attribute("asset", asset.to_string())
-            .add_attribute("recipient", recipient.to_string())
-            .add_attribute("generator_address", self.generator_addr.to_string());
-
-        Ok(Response::new()
-            .add_submessage(SubMsg {
-                id: ASTROPORT_LOCK_TOKENS_REPLY_ID,
-                msg: stake_msg,
-                gas_limit: None,
-                reply_on: ReplyOn::Success,
-            })
-            .add_event(event))
-    }
-
-    fn unstake(&self, _deps: Deps, asset: Asset, recipient: Addr) -> Result<Response, CwDexError> {
-        let unstake_msg = CosmosMsg::Wasm(
-            (WasmMsg::Execute {
-                contract_addr: self.lp_token_addr.to_string(),
-                msg: to_binary(&GeneratorExecuteMsg::Withdraw {
-                    lp_token: self.lp_token_addr.to_string(),
-                    amount: asset.amount,
-                })?,
-                funds: vec![],
-            }),
-        );
-
-        let event = Event::new("apollo/cw-dex/unstake")
-            .add_attribute("type", "astroport_staking")
-            .add_attribute("recipient", recipient.to_string());
-
-        Ok(Response::new().add_message(unstake_msg).add_event(event))
-    }
-
-    fn claim_rewards(&self, _recipient: Addr) -> Result<Response, CwDexError> {
-        let claim_rewards_msg = CosmosMsg::Wasm(
-            (WasmMsg::Execute {
-                contract_addr: self.generator_addr.to_string(),
-                msg: to_binary(&GeneratorExecuteMsg::ClaimRewards {
-                    lp_tokens: vec![self.lp_token_addr.to_string()],
-                })?,
-                funds: vec![],
-            }),
-        );
-
-        let event =
-            Event::new("apollo/cw-dex/claim_rewards").add_attribute("type", "astroport_staking");
-        Ok(Response::new().add_message(claim_rewards_msg).add_event(event))
-    }
-}
-
-/// This enum describes a Terra asset (native or CW20).
-#[cw_serde]
-pub struct AstroStableSwapAsset {
-    /// Information about an asset stored in a [`AssetInfo`] struct
-    pub info: AstroAssetInfo,
-    /// A token amount
-    pub amount: Uint128,
-    /// decimal precision
-    pub precision: u8,
-}
