+++ src/implementations/astroport/mod.rs
@@ -2,6 +2,8 @@ mod base_pool;
 mod helpers;
 mod staking;
 mod xyk;
-
+mod stableswap;
 pub use staking::*;
 pub use xyk::*;
+mod querier;
+pub use stableswap::*;
